--- src/main/java/codesquad/team05/domain/accommodation/accommodation_hashtag/AccommodationHashtagRepository.java
@@ -1,6 +1,0 @@
-package codesquad.team05.domain.accommodation.accommodation_hashtag;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccommodationHashtagRepository extends JpaRepository<AccommodationHashtag, Long> {
-}
