--- src/main/java/codesquad/team05/domain/accommodation/like/LikeRepository.java
@@ -1,6 +1,0 @@
-package codesquad.team05.domain.accommodation.like;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface LikeRepository extends JpaRepository<Like, Long> {
-}
