--- src/main/java/codesquad/team05/domain/accommodation/reservation/review/Review.java
@@ -1,34 +1,0 @@
-package codesquad.team05.domain.accommodation.reservation.review;
-
-import codesquad.team05.domain.accommodation.Accommodation;
-import codesquad.team05.domain.user.User;
-import jakarta.persistence.*;
-import lombok.Setter;
-
-@Entity
-@Setter
-@Table(name = "reviews")
-public class Review {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private int rating;
-    private String comment;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "accommodation_id")
-    Accommodation accommodation;
-
-    public void setUser(User user) {
-        this.user = user;
-    }
-
-    public void setAccommodation(Accommodation accommodation) {
-        this.accommodation = accommodation;
-    }
-}
