--- src/main/java/codesquad/team05/domain/accommodation/reservation/review/ReviewRepository.java
@@ -1,6 +1,0 @@
-package codesquad.team05.domain.accommodation.reservation.review;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReviewRepository extends JpaRepository<Review, Long> {
-}
