--- src/main/java/codesquad/team05/domain/exception/PersonCountExceededException.java
@@ -1,14 +1,0 @@
-package codesquad.team05.domain.exception;
-
-public class PersonCountExceededException extends RuntimeException {
-    public PersonCountExceededException() {
-    }
-
-    public PersonCountExceededException(String message) {
-        super(message);
-    }
-
-    public PersonCountExceededException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
