+++ src/main/java/codesquad/team05/domain/like/Like.java
@@ -1,0 +1,22 @@
+package codesquad.team05.domain.like;
+
+import codesquad.team05.domain.accommodation.Accommodation;
+import codesquad.team05.domain.user.User;
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "likes")
+public class Like {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "accommodation_id", nullable = false)
+    private Accommodation accommodation;
+}
