+++ src/main/java/codesquad/team05/domain/like/LikeRepository.java
@@ -1,0 +1,6 @@
+package codesquad.team05.domain.like;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LikeRepository extends JpaRepository<Like, Long> {
+}
