+++ src/main/java/codesquad/team05/domain/reservation/ReservationRepository.java
@@ -1,0 +1,10 @@
+package codesquad.team05.domain.reservation;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+
+    List<Reservation> findReservationByAccommodationId(Long accommodationId);
+}
