+++ src/main/java/codesquad/team05/domain/review/ReviewRepository.java
@@ -1,0 +1,6 @@
+package codesquad.team05.domain.review;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+}
