+++ src/main/java/codesquad/team05/service/AccommodationService.java
@@ -3,8 +3,8 @@ package codesquad.team05.service;
 import codesquad.team05.domain.accommodation.Accommodation;
 import codesquad.team05.domain.accommodation.AccommodationRepository;
 import codesquad.team05.domain.picture.Picture;
-import codesquad.team05.web.dto.request.accommodation.AccommodationSave;
-import codesquad.team05.web.dto.request.accommodation.AccommodationUpdate;
+import codesquad.team05.web.accommodation.dto.request.AccommodationSave;
+import codesquad.team05.web.accommodation.dto.request.AccommodationUpdate;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import lombok.RequiredArgsConstructor;
