+++ src/main/java/codesquad/team05/web/accommodation/dto/response/AccommodationResponse.java
@@ -1,0 +1,21 @@
+package codesquad.team05.web.accommodation.dto.response;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class AccommodationResponse {
+
+    private Long id;
+    private String name;
+    private int price;
+    private String address;
+    private int maxCapacity;
+    private int roomCount;
+    private int bedCount;
+    private String description;
+    private String amenity;
+    private List<PictureDto> pictures;
+
+}
