+++ src/main/java/codesquad/team05/web/accommodation/dto/response/PictureDto.java
@@ -1,0 +1,9 @@
+package codesquad.team05.web.accommodation.dto.response;
+
+import lombok.Data;
+
+@Data
+public class PictureDto {
+    private String url;
+
+}
