--- src/main/java/codesquad/team05/web/controller/UserController.java
@@ -1,4 +1,0 @@
-package codesquad.team05.web.controller;
-
-public class UserController {
-}
