--- src/main/java/codesquad/team05/web/dto/request/reservation/ReservationUpdate.java
@@ -1,28 +1,0 @@
-package codesquad.team05.web.dto.request.reservation;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-@AllArgsConstructor
-public class ReservationUpdate {
-
-    @NotBlank
-    private int amount;
-
-    @NotBlank
-    private int personCount;
-
-    @NotBlank
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
-    private LocalDate checkIn;
-
-    @NotBlank
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
-    private LocalDate checkOut;
-
-}
