--- src/main/java/codesquad/team05/web/dto/response/accommodation/PictureDto.java
@@ -1,9 +1,0 @@
-package codesquad.team05.web.dto.response.accommodation;
-
-import lombok.Data;
-
-@Data
-public class PictureDto {
-    private String url;
-
-}
