--- src/main/java/codesquad/team05/web/dto/response/user/UserResponse.java
@@ -1,12 +1,0 @@
-package codesquad.team05.web.dto.response.user;
-
-import lombok.Data;
-
-@Data
-public class UserResponse {
-
-    private Long id;
-    private String loginId;
-    private String name;
-
-}
