+++ src/main/java/codesquad/team05/web/reservation/dto/request/ReservationUpdate.java
@@ -1,0 +1,28 @@
+package codesquad.team05.web.reservation.dto.request;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.time.LocalDate;
+
+@Data
+@AllArgsConstructor
+public class ReservationUpdate {
+
+    @NotBlank
+    private int amount;
+
+    @NotBlank
+    private int personCount;
+
+    @NotBlank
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
+    private LocalDate checkIn;
+
+    @NotBlank
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
+    private LocalDate checkOut;
+
+}
