+++ src/main/java/codesquad/team05/web/user/dto/response/UserResponse.java
@@ -1,0 +1,12 @@
+package codesquad.team05.web.user.dto.response;
+
+import lombok.Data;
+
+@Data
+public class UserResponse {
+
+    private Long id;
+    private String loginId;
+    private String name;
+
+}
